Adding Dashboard 'real-time' mode (refresh based), including experimental motion detection
Adding Dashboard 'real-time' mode (refresh based) monitoring newly activated lights, motion detection, temperatures, and alerts for unreachable or low-battery devices. From trial and error, I feel the motion detect tracking is somewhat experimental at this stage as I've had cases where it was reported and others when it wasn't. This could be down to Hue updating the API in timely fashion or not. I will keep an eye and improve it if possible in upcoming fixes.